requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
click==8.1.7
dj-database-url==2.1.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
uvicorn==0.27.1
